Define blacksmith commands on its own file
Summary: This is a preparation for including `blackmisth-containerized-build-command` in this repository

Reviewers: jfuentes

Differential Revision: http://phabricator.bitnami.com:8080/D13404
